Extended error handling for email
If emailjs can't find any of the ids, now gives an error instead of
waiting forever. This is a temporary fix while I figure out how to get
the env variables into the docker project. Will likely have to start
work on creating an express backend or similar to be able to do a PHP
email server thing